CARRY: HACK: insert worker VMs into aro load balancer backend
This ensures that worker VMs are always in the aro load balancer backend, and
are thus always eligible for the associated outbound rule, and thus should have
internet access.  As a result a fake inbound service is not necessary.

--

Original commit: https://github.com/openshift/installer/commit/638b55246e6478e20c250a13cb034d69708e8845

We no longer need the hunk setting PublicLoadBalancer bit as the installer
already does it for clusters with OutboundType=Loadbalancer (applies to ARO).
The remaining PR is to remove dummy outbound service from the installer as it
seem to be no longer needed for everyone on Azure.

PR: https://github.com/openshift/installer/pull/4880